Update functions.R
Made some changes to the documentation wording and deleted some commented code which is not necessary. Also, the example code is slightly changed (set tau_main = 0.4 instead of 0.7) to ensure that the example returns a positive p value = 0.1 (instead of 0 with the original tau_main), which show the tendency to reject the null.