Greedy algorithm and erroneous decomposition
Add greedy algorithm as a method to decompose the NEgraph. Also add an error message to the final output of the clique_test function if for some focal assignments in the a biclique that contains Zobs, exposures for each unit are the same. Slightly modified the given greedy algorithm to get the true decomposition of NEgraph.